update to allow calling api calls with apiUrl and apiKey
