+++ package.json
@@ -53,7 +53,6 @@
     "@types/jest": "^29.5.12",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
-    "axios": "1.x",
     "cosmjs-types": "0.9.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
