+++ packages/client/package.json
@@ -63,7 +63,6 @@
     "@injectivelabs/sdk-ts": "1.15.3",
     "@keplr-wallet/unit": "^0.12.143",
     "@solana/wallet-adapter-base": "^0.9.23",
-    "axios": "1.x",
     "bech32": "^2.0.0",
     "cosmjs-types": "^0.9.0",
     "keccak": "3.0.4"
