+++ packages/client/src/public-functions/setApiOptions.ts
@@ -1,11 +1,11 @@
-import { ClientState, SkipApiOptions } from "../state";
+import { Fetch, SkipApiOptions } from "src/utils/fetchClient";
 import { createRequestClient } from "../utils/generateApi";
 
 export const setApiOptions = (options: SkipApiOptions = {}) => {
-  ClientState.requestClient = createRequestClient({
+  Fetch.client = createRequestClient({
     baseUrl: options.apiUrl || "https://api.skip.build",
     apiKey: options.apiKey,
   });
 
-  ClientState.setClientInitialized();
+  Fetch.setClientInitialized();
 };
