+++ packages/client/src/state.ts
@@ -1,16 +1,15 @@
 import { GeneratedType, Registry } from "@cosmjs/proto-signing";
 import { AminoConverters, AminoTypes, SigningStargateClient } from "@cosmjs/stargate";
 import { Chain, Asset, ChainAffiliates } from "./types/swaggerTypes";
-import { ApiResponse, createRequestClient } from "./utils/generateApi";
+import { ApiResponse } from "./utils/generateApi";
 import { getMainnetAndTestnetChains } from "./private-functions/getMainnetAndTestnetChains";
 import { getMainnetAndTestnetAssets } from "./private-functions/getMainnetAndTestnetAssets";
 import { balances } from "./api/postBalances";
 import { EndpointOptions, SignerGetters, ValidateGasResult } from "src/types/client-types";
+import { SkipApiOptions } from "./utils/fetchClient";
 
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
 export class ClientState {
-  static requestClient: ReturnType<typeof createRequestClient>;
-
   static aminoTypes: AminoTypes;
   static registry: Registry;
 
@@ -62,26 +61,6 @@ export class ClientState {
 
   static signingStargateClientByChainId: Record<string, SigningStargateClient> = {};
   static validateGasResults: ValidateGasResult[] | undefined;
-
-  private static isInitialized = false;
-  private static resolveInitialization: () => void;
-  static clientInitialized: Promise<void> = new Promise<void>((resolve) => {
-    this.resolveInitialization = () => {
-      if (!this.isInitialized) {
-        this.isInitialized = true;
-        resolve();
-      }
-    };
-  });
-
-  static setClientInitialized() {
-    this.resolveInitialization();
-  }
-}
-
-export type SkipApiOptions = {
-  apiUrl?: string;
-  apiKey?: string;
 }
 
 export type SkipClientOptions = SkipApiOptions & {
