+++ packages/widget/src/state/skipClient.ts
@@ -64,6 +64,7 @@ export const skipAssetsAtom = atomWithQuery((get) => {
         includeCw20Assets: true,
         includeSvmAssets: true,
         onlyTestnets,
+        abortDuplicateRequests: true,
       });
 
       return flattenData(response as Record<string, Asset[]>, chains.data);
@@ -83,6 +84,7 @@ export const skipChainsAtom = atomWithQuery((get) => {
         includeEvm: true,
         includeSvm: true,
         onlyTestnets,
+        abortDuplicateRequests: true,
       });
       return response;
     },
